Fix for compile issue (pc/sim)

+++ Tactility/Private/Tactility/app/wifimanage/View.h
@@ -26,7 +26,7 @@ class View final {
     void updateScanning();
     void updateNetworkList();
     void updateConnectToHidden();
-    void createSsidListItem(const service::wifi::ApRecord& record, bool isConnecting, int index);
+    void createSsidListItem(const service::wifi::ApRecord& record, bool isConnecting, size_t index);
 
     static void showDetails(lv_event_t* event);
     static void connect(lv_event_t* event);
